Bug: #1741
Found two bugs in the code, one that the ServerPartitionMonitor code that checked for changes in a partition was not checking the duplicate sop policy (it also was not checking on the new fields related to the study integrity queue).  The second problem was that the scp extensions were caching the partition object, and not the context object, and thus not catching changes in the partition.  Updated the SCP code to cache the context object, which is updated when the partition changes.