Bug: #2105
Fixed the XML schema mechanism used for validating server rules so that the schema generated is conditional on the rule type.  We now use generics with the XML compiler to get a context to pass to the schema generation routine for each action operator.  This is then used to return null, or the valid schema information, if applicable to the rule type.