Bug: #2105
Fixed bug from yesterday's checkin on this.  Made a bad assumption that server side validation was not done, and had some bad code because of it.  Turned off the strict validation on the server side, since it couldn't be done with the current control.  The validation is properly done now through a web service, so there's no loss in functionality.