Bug: #3113
Partial commit.  Updated the database to remove StudyStatusEnum / QueueStudyStateEnum from the Study table, this is now only in the StudyStorage table.  Made appropriate changes in the code to make it compile with this, and the stored procedures to basically work with this.  Still need to go and reexamine all the places where this is used to make sure its implemented properly and the new locking scheme is working properly.