Bug: #3113
Partial commit.  The ServiceLock services were not locking properly, fixed this.  Also, there was a typo or two in the stored procedures that was fixed.  Also added the queue study state to be displayed on the StudyStorage tab in study details.  Also fixed an "oops" with the restore functionality and releasing the lock.  it worked with nearline studies before, but not when restoring over the top of an online study.