Ticket: #3335 Draft OB-US structured report is not visible in biography
Added some code into the ob-us-mpps.htm page to allow it to be printed directly from the browser with patient info inserted at top of page.  This was a quick and dirty hack and doesn't look very good right now.  Perhaps it can be cleaned up a bit but probably not too much.