Ticket: #3470 Reporting preview, procedures sometimes show blank status
If the procedure is any status other than IP or CM, the procedure status itself is now shown (not the "reporting" status).  If the procedure is IP but not yet fully performed, the status will appear blank.  The status will also appear blank if the procedure has been linked to the report of another procedure.  This is ambiguity is not ideal but it cannot be fixed without server-side changes.