Ticket: #2888
Added new mechanism to save log messages into the database, display them in the database, and then archive them to one of the filesystems after a configurable number of days.  We take advantage of the AdoAppender in log4net to insert into the database, and then use a ServiceLock process for later removing the log entries from the database after a configured number of days.  Also refactored the DataSource locations in the web projects and cleaned up the thread naming mechanism a bit in the dicom library and the thread pools to make the logs a bit cleaner.