Bug: #3855
This problem arose from a message having private tags 0031,0002 , 0031,0003 , and 0031,0004 in a message.  We were assuming the tags were LO Vr, when in reality they were just standard tags with binary values.  Our assumption of LO VR caused us to corrupt our outgoing data when we wrote them out to disk, which caused the actual problem reported in the bug.