Ticket: #2334
Added upgrade scripts from 1.35 -> Enterprise.  Note that RC3 of 1.35 is taken as the version to upgrade from.  This will have to be updated if there's a RC4...  Also decided to change the enumerated values again as per Ticket: #3809.  We had several descriptions greater than 256 characters, so just made the longdescription on all enumerated value tables 512 instead of 256.