Ticket #3941: Study w/ Invalid Birth Date goes into infinite loop with reconcilliation
Fixed comparison code in StudyComparer. Use DicomAttrubyte.ToString() to retrieve the value in the dicom message for comparison. The same method is used when the database is populated.

Modified RenameFileCommand to backup and restore the source file when rolling back.

Added OpValidationCommand to the validate the result of the operation. It checks if the file has been copied properly to the Reconcile folder and triggers restore of the files.

Added method to simulate errors/exceptions. In DEBUG mode, errors are simulated at the end of RenameFileCommand and MoveReconcileImageCommand.

Modified StudyProcessItemProcessor to log the information about the SOP that is being reconciled, and list the problems that causes the reconciliation.