Ticket #3941: Study w/ Invalid Birth Date goes into infinite loop with reconcilliation
Refactored the code to prevent infinite loop reconciliation. Moved classes that were specific to Study Reconcile processor into Common and added StudyHelper.Compare() which uses StudyComparer to compare a dicom file against a study. The comparison is done when the images are processed, as well as when the ReconcileStudy entry is processed (ie, images in the Reconcile folder are updated but not yet copied to the study folder). If difference(s) is found, the ReconcileStudy entry will fail and the images are left in the Reconcile folder instead.