Ticket: #2334
Refactored the DB upgrade mechanism so the base upgrade classes are now in Enterprise.Core and can be used by the RIS team.  This included a breaking change to IPersistentStore, where we added a version property for the persistent store.  The RIS project has been updated so it doesn't cause problems.  Also did some refactoring of the code to make naming conventions proper.

For the ImageServer, the new command line to run the upgrade is:

ClearCanvas.ImageServer.Executable ClearCanvas.Enterprise.Core.Upgrade.PersistentStoreUpgradeApplication