Requirement: #3808
Added support for viewing the study details page from the ArchiveQueue and RestoreQueue pages.  There is now a button on both pages to view details.  Also made small cleanup to the study details page, specifically tot he Archive tab, the archive study storage information was scrolling off the screen to the right, and not visible, added a table to force it to be below the archive queue.