Ticket: #3407 Add Audit trail logging to Web UI.
Interim commit.
1. Removed unused auditing classes from ClearCanvas.Common.Auditing.
2. Added some new classes that have a similar but different architecture, which essentially provide a convenient "client" class, rather than having client code use the service directly.
3. Added the IAuditService interface and corresponding implementation.

Client code should not use the IAuditService directly.  Rather, create an instance of an AuditLog (in ClearCanvas.Common.Audit) and call it's WriteEntry method.  If the Enterprise.Common plugin is loaded, this will automatically be directed to the IAuditService.

Outstanding: documentation and converting some of the existing RIS code to take advantage of the new code.