Bug: #4251 Bug: #3720
Added new logging for each of the WorkQueue entries to signal when we start working on the queue entry, and the Patient Name / Id / A# / Study UID of the study being worked on.  Also at the same time cleaned up the CanStart() method used in the WorkQueue entries.  This method is now responsible for postponing the queue entry, if required.  Also moved the handling of loading the StorageLocation into the BaseItemProcessor, since this is done by all the workqueue processors.  Also cleaned up a small warning during builds.