Ticket: #1013
Add support for further Normalized messages based on a patch submitted by Dan Blanchard here:

http://www.clearcanvas.ca/dnn/Community/Forums/tabid/69/afv/topic/aff/9/aft/13260/Default.aspx

This adds normalized sending of response messages.   Note that I believe N-EVENT messages are the last type of N messages that need to be developed.

Also added a new Platform.IsLogLevelEnabled method for determining if a logging level is enabled.  As part of this patch, Dan submitted a mechanism to log the contents of messages send/received at the Debug level.  This code needed to check if the debug level was enabled before doing the log.