Bug: #3636
Added Alerts when the DICOM listener cannot be started.  Had to make some small changes in the DICOM API to return false when failing to startup, and to do the initial listening from the base thread.

Also cleaned up a bug in the Alert page of the Web GUI, we would only ever display 25 results on the page.  Also, added sorting of the results with the most recent value sorted first.