Bug: #4225
Found issue with queries to Fusion where on the last C-MOVE-RSP message, the counts for pending operations were not being set, which was causing us to think no images moves, and failing incorrectly.  Also added configuration options for move timeouts, and backed out previous checkin that had repeated timeouts.  We can now set the timeout in the config file.