Bug: #4196
Made some sweeping performance improvements.  Found some areas when doing queries where we were doing multiple queries per row, and were opening separate connections per query.  Reworked this to make the queries more efficient.  Also changed the code generator so the Load() method on each entity object uses IPersistenceContext instead of IReadContext, which increased the size of this commit.  Also, backed off on the frequency of some of our polling of the queues to reduce db stress a bit.  it was excessive before.