Ticket: #4294 Allow users to create own worklists
Changed the worklist ownership model so that it supports both Staff and StaffGroup ownership, similar to the canned-text model.  This was necessary to avoid issues with managing group subscriptions to a staff-owned worklist.  Now a staff owned worklist can only be subscribed to a single staff and no groups, and a group-owned worklist can only be subscribed to a single group, and no staff.