Bug: #4275
Partial commit.  Added saving of the DICOM files to a temporary file, then moving to the final name, to reduce the change of corruption of the files.  This was applied to all the places we're saving a dicom file.  Still need to add checks on counts for the study after processing is done.