Bug: #4300
First attempt at addressing issue with multiple stored procedure changes.  Changed the query for rows to work on routines to use UPDLOCK forcing the rows to be locked, after the query finds them, which should reduce contention.  Also updated the InsertWorkQueue procedure to use NOLOCK, in hopes of not seeing these locks.