Bug: #3665
Reworked how the study level rules are handled for both after archival and when a study is complete processing.  We now use the study xml to check for the first non-compressed image in each series to apply the rules against, if all the images are compressed, then the rules are applied to the first image in each series.  This should address the problem.