Bug: #4391
Cleanup of directory handling in general.  Found bug with tier migration, where we were not cleaning up the source date directory, if there were no files in it.  Also added deletion of empty directories in the ServiceLock entries that are traversing the directories.  Finally, did some general code cleanup and documentation.