Ticket: #4318
Added simple mechanism that checks if there is available memory before queueing up workqueue entries for processing.  Acts in the same way as if the threads are used up.  Also added an alert if the memory limit is exceeded for more than 60 minutes.  as a default, 256MB of RAM must be available to launch a workqueue thread.  Note that the check is disabled if the configuration option is set to 0.