Bug: #4420: Header Streaming service returns "Could not find file xxxxxxx.gz" while study is being compressed.
Added study status check before attempting to streaming the header or the image back the the client. Basically users will be blocked from accessing the study while it is being used. An error message will be returned as part of the FaultException (for the header streaming) or a Http Status Description field (for image streaming). If the study is near-line, the message also indicates the study is being restored by the server.