Bug: #4420: Header Streaming service returns "Could not find file xxxxxxx.gz" while study is being compressed.
Fixed the address where the http server is listening at. If the server has multiple IP addresses, using the exact IP address will prevent the server from accepting requests from clients which use the other IP addresses. Changed the HttpListenerShred class to accept the port and path of the url and use "+" in place of the ip address when instantiate the listener. This instructs the listener to accept all requests sent to the given port, regardless of the IP address in the requested URI.