Bug: #4179
Worked on handling of Excluded tags to make them more efficient.  Turns out the ticket was a bit misleading, the excluded tags were being dealt with in the base instance properly in the code.  It was just when a tag was not in the base instance, but was in the derived instances that the tags were added, which was appropriate.

Additionally added in some more code to handle this better.  The previous code blocked out all OB/OW/OF tags, and they would not necessarily end up in the excluded tags, we now include these.  There was a small hole if the excluded attribute was in the base instance, but the other instances included the tag, this did not work properly.

Don't have any specific studies for this, so this should just be regression tested.