Ticket: #4500 Settings group meta-data must be imported to configuration store
Moved the Configuration setup application out of Enterprise.Configuration and into Enterprise.Commmon. It is now a generic setup application that can be run with any server install - if the server being installed is not the enterprise server, it will connect to the enterprise server to import the settings groups.