Bug: #4504
Created a new FileUtils.Delete() method that retrieves deletes after a configurable delay if an IOException occurs during the delete.  Retrieves are done until a timeout occurs, after which time the exception is still thrown.

Did a code inspection for compression with how we were dealing with the source file.  Both with how individual frames were loaded when we were storing pixel data references, and how the file was orginally opened.  In all cases it appears that the the opening of the file has a "using()" block that loads the file correctly, and is disposed after the file is closed.  Based on this, decided to implement the retry.  Apllied this across the library in a number of areas that appeared might have the issue, in addition to the general case of compression seen on the test servers.