Requirement: #4447
Added support for compressing images when they're received.  Added new Server rule type, SopCompressed, which is for compressing specific SOP Instances.  This rule type also must be applied at SopProcessed time.  Made corresponding DB upgrade script changes for this.  Also adjusted the processing of the objects to include this update. We now also will update the StudyStorage status and transfer syntax for the post compression state.  Also added sample rules for the new rule types.  Made updates to the OpenJpeg Jpeg2000 code, even though we're not including this in the official builds.  Also made a small adjustment to the server rules dialog to make it wider.  Finally, made some adjustments to the rules engine so you cna either define only the types of rules to include, or you can define the types of rules to omit when the rules engine is built.