Ticket: #4523 Develop Auto-Recovery Mechanism for WorkQueue Failures
Interim commit. Added auto-recovery to study compression and tier-migration. If the entry failed because of mismatch number of instances in the xml and db, the entry will be deleted and replaced with a ReprocessStudy entry. All rules will be re-applied when the study is reprocessed. Corrupted images will be deleted when they are encountered during the reprocessing and alerts will be generated.

Other changes:

- Added a setting to simulate file corruption when receiving images. Disabled by default.
- Updated the work queue failure alerts to include more information in the message.