Bug: #4693
Updates to the Study Reprocess and Study Reinventory ServiceLock entries to deal with this properly.  Both were just using the directory name as the study instance uid.  made an update if the study is not found this way, we actually go to the first file found, and use the study instance uid from this file and re-lookup the study.