Ticket: #4523 Develop Auto-Recovery Mechanism for WorkQueue Failures
Interim commit. Changed how the auto-recovery works with the fix made for ticket #4677. ReprocessStudy no longer removes existing work queue and SIQ entries.

Validation is now done at the end of the processing before the entry is completed. Instead of triggering a reprocessing all the time, we first verify the study xml against the filesystem. If the number of instances in the study xml matches number of files in the filesystem, simply update the number of instances in the db. Otherwise, we insert a ReprocessStudy entry.  ReprocessStudy is always processed before any other entries. Once the study has been reprocessed, the current entry can then resume (and complete).

When reprocessing is required, a history record will be inserted to indicate what triggered it and reason it needed to be reprocessed.

- Added couple of stored procedures to update the series/instance counts in the study and series tables. These procedures will automatically update the counts in the Patient table according to the differences.
- Updated CompressStudyItemProcessor to skip SOP which is not in the study (auto-recovery may remove SOPs that are not longer valid).
- Enabled automatic recovery for ReconcileStudy, ProcessStudy.