Bug: #4703
Refined the logging a bit for archiving of the application log and purging of alerts in hopes of finding out more information if this is a real problem.  We now have an explicit log line when we assign a filesystem to the ServiceLock initially.  We also have a bit clearer logs as to what is being flushed out when the logs are being purged.