Ticket: #4843 - Failure seen related to paths when auto-reconciling and when filesystem path is long
Schema/Configuration Changes: None

Removed the study folder structure in the backup directory. Also made the code to delete the original folder when the command is disposed to simplify the rollback. Instead of deleting the folder after updating the files and then restoring it by copying the files in the backup folder back to the original folder, a flag is added to indicate whether the original should be deleted and it is set to false if the filesystem needs to be restored.